paletteer-package {paletteer}	R Documentation

paletteer: Comprehensive Collection of Color Palettes

Description

The choices of color palettes in R can be quite overwhelming with palettes spread over many packages with many different API's. This packages aims to collect all color palettes across the R ecosystem under the same package with a streamlined API.

Author(s)

Maintainer: Emil Hvitfeldt emilhhvitfeldt@gmail.com

Authors:

See AUTHORS file.

See Also

Useful links:

https://emilhvitfeldt.github.io/paletteer/

https://github.com/EmilHvitfeldt/paletteer

Report bugs at https://github.com/EmilHvitfeldt/paletteer/issues

[Package paletteer version 1.7.0 Index]