paletteer_c {paletteer}	R Documentation

Get continuous palette by package and name

Description

Available package/palette combinations are available in the data.frame palettes_c_names.

Usage

paletteer_c(palette, n, direction = 1, call = caller_env(0))

Arguments

palette
Name of palette as a string. Must be on the form packagename::palettename.

n
Number of colors desired. Must be supplied.

direction
Either 1 or -1. If -1 the palette will be reversed.

call
The execution environment of a currently running function, e.g. caller_env(). The function will be mentioned in error messages as the source of the error. See the call argument of abort() for more information.

Value

A vector of colors.

Examples

paletteer_c("scico::berlin", 100)

[Package paletteer version 1.7.0 Index]