Get discrete palette by package and name

Available package/palette combinations are available in the data.frame palettes_d_names. Both package and palette can be supplied as symbols or strings.

paletteer_d(package, palette, n, direction = 1, type = c("discrete",
  "continuous"))

Arguments

package
Name of package from which the palette is desired as string or symbol.

palette
Name of palette as string or symbol.

n
Number of colors desired. If omitted, returns complete palette.

direction
Either 1 or -1. If -1 the palette will be reversed.

type
Either "discrete" or "continuous". Colors are interpolated if "continuous" is picked. Defaults to "discrete".

Value

A vector of colors.

Examples

paletteer_d("nord", "frost")

#> [1] "#8FBCBB" "#88C0D0" "#81A1C1" "#5E81AC"

paletteer_d("wesanderson", "Royal1", 3)

#> [1] "#899DA4" "#C93312" "#FAEFD1"

paletteer_d("Redmonder", "dPBIPuOr", 14, type = "continuous")

#>  [1] "#53354D" "#73496A" "#9C6D92" "#BD90B3" "#CBA7C3" "#D8BED2" "#EBE1E9"
#>  [8] "#F8E8E1" "#FED1BD" "#FEC1A5" "#FEB18E" "#E0906B" "#B06846" "#7F4B33"