palettes_d {paletteer}	R Documentation

Complete list of fixed discrete palettes

Description

A list of all the discrete palettes of fixed lengths included in this package. Structured as a list of lists with the first level being each package and the second level being the palettes available in that package.

Usage

palettes_d

Format

A list of 73 lists.

Details

Contains palettes from the following packages: ButterflyColors, DresdenColor, IslamicArt, LaCroixColoR, Manu, MapPalettes, MetBrewer, MexBrewer, MoMAColors, NatParksPalettes, NineteenEightyR, PNWColors, Polychrome, PrettyCols, RColorBrewer, RSkittleBrewer, Rdune, Redmonder, amerika, awtools, basetheme, beyonce, blueycolors, calecopal, colRoz, colorBlindness, colorblindr, dichromat, dutchmasters, feathers, fishualize, futurevisions, ggpomological, ggprism, ggsci, ggthemes, ggthemr, ghibli, grDevices, impressionist.colors, jcolors, khroma, lisa, ltc, miscpalettes, musculusColors, nationalparkcolors, nbapalettes, nord, ochRe, palettesForR, palettetown, pals, peRReo, poisonfrogs, rcartocolor, rockthemes, rtist, severance, soilpalettes, suffrager, tayloRswift, tidyquant, trekcolors, tvthemes, unikn, vangogh, vapeplot, vapoRwave, waRhol, werpals, wesanderson, yarrr.

data.frame of palette names can be found here palettes_d_names.

[Package paletteer version 1.7.0 Index]