palettes_d_names {paletteer}	R Documentation

Data Frame of all fixed discrete palettes

Description

A data.frame of the names and other characteristics of the discrete palettes of fixed lengths included in this package.

Usage

palettes_d_names

Format

A data.frame of 2545 observations with 4 variables:

package

character, name of package

palette

character, name of palette

length

integer, number of colors in palette

type

character, type of palette

Details

Contains palettes names from the following packages: ButterflyColors, DresdenColor, IslamicArt, LaCroixColoR, Manu, MapPalettes, MetBrewer, MexBrewer, MoMAColors, NatParksPalettes, NineteenEightyR, PNWColors, Polychrome, PrettyCols, RColorBrewer, RSkittleBrewer, Rdune, Redmonder, amerika, awtools, basetheme, beyonce, blueycolors, calecopal, colRoz, colorBlindness, colorblindr, dichromat, dutchmasters, feathers, fishualize, futurevisions, ggpomological, ggprism, ggsci, ggthemes, ggthemr, ghibli, grDevices, impressionist.colors, jcolors, khroma, lisa, ltc, miscpalettes, musculusColors, nationalparkcolors, nbapalettes, nord, ochRe, palettesForR, palettetown, pals, peRReo, poisonfrogs, rcartocolor, rockthemes, rtist, severance, soilpalettes, suffrager, tayloRswift, tidyquant, trekcolors, tvthemes, unikn, vangogh, vapeplot, vapoRwave, waRhol, werpals, wesanderson, yarrr.

Full list of palette can be found here palettes_d.

[Package paletteer version 1.7.0 Index]