palettes_dynamic {paletteer}	R Documentation

Complete list of dynamic palettes

Description

A list of all the dynamic palettes included in this package. Structured as a list of lists with the first level being each package and the second level being the palettes available in that package, and the third level being the palettes for various lengths.

Usage

palettes_dynamic

Format

A list of 3 lists.

Details

Contains palettes from the following packages: cartography and ggthemes.

Full list of palette names can be found here palettes_dynamic_names.

[Package paletteer version 1.7.0 Index]