palettes_dynamic_names {paletteer}	R Documentation

Data Frame of all fixed discrete palettes

Description

A data.frame of the names and other characteristics of the dynamic palettes included in this package.

Usage

palettes_dynamic_names

Format

A data.frame of 25 observations with 4 variables:

package

character, name of package

palette

character, name of palette

length

integer, maximal number of colors in palette

type

character, type of palette

Details

Contains palettes from the following packages: cartography and ggthemes.

Full list of palette can be found here palettes_dynamic.

[Package paletteer version 1.7.0 Index]