paletteer: Comprehensive Collection of Color Palettes

The choices of color palettes in R can be quite overwhelming with palettes spread over many packages with many different API's. This packages aims to collect all color palettes across the R ecosystem under the same package with a streamlined API.

See also

Author

Maintainer: Emil Hvitfeldt emilhhvitfeldt@gmail.com

Authors:

See AUTHORS file.