Complete list of fixed discrete palettes

A list of all the discrete palettes of fixed lengths included in this package. Structured as a list of lists with the first level being each package and the second level being the palettes available in that package.

palettes_d

Format

A list of 44 lists.

Details

Contains palettes from the following packages: awtools, basetheme, calecopal, colorblindr, colRoz, dichromat, dutchmasters, DresdenColor, fishualize, futurevisions, ggsci, ggpomological, ggthemes, ggthemr, ghibli, grDevices, IslamicArt, jcolors, LaCroixColoR, lisa, nationalparkcolors, NineteenEightyR, nord, ochRe, pals, palettesForR_pals, Polychrome, MapPalettes, miscpalettes, PNWColors, rcartocolor, RColorBrewer, Redmonder, RSkittleBrewer, tidyquant, trekcolors, tvthemes, unikn, vapeplot, vapoRwave, werpals, wesanderson and yarrr.

data.frame of palette names can be found here palettes_d_names.